+++ internal/sdkprovider/userconfig/serviceintegration/external_opensearch_logs.go
@@ -16,9 +16,9 @@ func externalOpensearchLogsUserConfig() *schema.Schema {
 		Elem: &schema.Resource{Schema: map[string]*schema.Schema{"selected_log_fields": {
 			Description: "The list of logging fields that will be sent to the integration logging service. The MESSAGE and timestamp fields are always sent.",
 			Elem: &schema.Schema{
-				Description:  "Enum: `HOSTNAME`, `PRIORITY`, `REALTIME_TIMESTAMP`, `service_name`, `SYSTEMD_UNIT`. Log field name.",
+				Description:  "Enum: `HOSTNAME`, `PRIORITY`, `REALTIME_TIMESTAMP`, `SYSTEMD_UNIT`, `service_name`. Log field name.",
 				Type:         schema.TypeString,
-				ValidateFunc: validation.StringInSlice([]string{"HOSTNAME", "PRIORITY", "REALTIME_TIMESTAMP", "service_name", "SYSTEMD_UNIT"}, false),
+				ValidateFunc: validation.StringInSlice([]string{"HOSTNAME", "PRIORITY", "REALTIME_TIMESTAMP", "SYSTEMD_UNIT", "service_name"}, false),
 			},
 			MaxItems: 5,
 			Optional: true,